Own classes for all windows.
