+++ src/BlockEntities/DropSpenserEntity.cpp
@@ -8,6 +8,7 @@
 #include "DropSpenserEntity.h"
 #include "../Entities/Player.h"
 #include "../Chunk.h"
+#include "../UI/DropSpenserWindow.h"
