+++ src/ClientHandle.cpp
@@ -12,6 +12,9 @@
 #include "BlockEntities/CommandBlockEntity.h"
 #include "BlockEntities/SignEntity.h"
 #include "UI/Window.h"
+#include "UI/AnvilWindow.h"
+#include "UI/BeaconWindow.h"
+#include "UI/EnchantingWindow.h"
 #include "Item.h"
 #include "Mobs/Monster.h"
 #include "ChatColor.h"
