+++ src/UI/Window.h
@@ -19,7 +19,6 @@ class cPlayer;
 class cWindowOwner;
 class cClientHandle;
 class cChestEntity;
-class cDropSpenserEntity;
 class cEnderChestEntity;
 class cFurnaceEntity;
 class cHopperEntity;
@@ -215,191 +214,6 @@ protected:
 	/// Distributes a_NumToEachSlot items into the slots specified in a_SlotNums; returns the total number of items distributed
 	int DistributeItemToSlots(cPlayer & a_Player, const cItem & a_Item, int a_NumToEachSlot, const cSlotNums & a_SlotNums);
 } ;  // tolua_export
-
-
-
-
-
-class cCraftingWindow :
-	public cWindow
-{
-	typedef cWindow super;
-public:
-	cCraftingWindow(int a_BlockX, int a_BlockY, int a_BlockZ);
-} ;
-
-
-
-
-
-class cAnvilWindow :
-	public cWindow
-{
-	typedef cWindow super;
-public:
-	cAnvilWindow(int a_BlockX, int a_BlockY, int a_BlockZ);
-
-	/** Gets the repaired item name. */
-	AString GetRepairedItemName(void) const { return m_RepairedItemName; }
-
-	/** Set the repaired item name. */
-	void SetRepairedItemName(const AString & a_Name, cPlayer * a_Player);
-
-	/** Gets the Position from the Anvil */
-	void GetBlockPos(int & a_PosX, int & a_PosY, int & a_PosZ);
-
-protected:
-	cSlotAreaAnvil * m_AnvilSlotArea;
-	AString m_RepairedItemName;
-	int m_BlockX, m_BlockY, m_BlockZ;
-} ;
-
-
-
-
-
-class cBeaconWindow :
-	public cWindow
-{
-	typedef cWindow super;
-public:
-	cBeaconWindow(int a_BlockX, int a_BlockY, int a_BlockZ, cBeaconEntity * a_Beacon);
-
-	cBeaconEntity * GetBeaconEntity(void) const { return m_Beacon; }
-
-	// cWindow Overrides:
-	virtual void OpenedByPlayer(cPlayer & a_Player) override;
-
-protected:
-	cBeaconEntity * m_Beacon;
-} ;
-
-
-
-
-
-class cEnchantingWindow :
-	public cWindow
-{
-	typedef cWindow super;
-public:
-	cEnchantingWindow(int a_BlockX, int a_BlockY, int a_BlockZ);
-	virtual void SetProperty(short a_Property, short a_Value, cPlayer & a_Player) override;
-	virtual void SetProperty(short a_Property, short a_Value) override;
-
-	/** Return the Value of a Property */
-	short GetPropertyValue(short a_Property);
-
-	cSlotArea * m_SlotArea;
-
-protected:
-	short m_PropertyValue[3];
-	int m_BlockX, m_BlockY, m_BlockZ;
-};
-
-
-
-
-
-class cFurnaceWindow :
-	public cWindow
-{
-	typedef cWindow super;
-public:
-	cFurnaceWindow(int a_BlockX, int a_BlockY, int a_BlockZ, cFurnaceEntity * a_Furnace);
-} ;
-
-
-
-
-
-class cDropSpenserWindow :
-	public cWindow
-{
-	typedef cWindow super;
-public:
-	cDropSpenserWindow(int a_BlockX, int a_BlockY, int a_BlockZ, cDropSpenserEntity * a_Dispenser);
-} ;
-
-
-
-
-
-class cHopperWindow :
-	public cWindow
-{
-	typedef cWindow super;
-public:
-	cHopperWindow(int a_BlockX, int a_BlockY, int a_BlockZ, cHopperEntity * a_Hopper);
-} ;
-
-
-
-
-
-class cChestWindow :
-	public cWindow
-{
-public:
-	cChestWindow(cChestEntity * a_Chest);
-	cChestWindow(cChestEntity * a_PrimaryChest, cChestEntity * a_SecondaryChest);
-	~cChestWindow();
-
-	virtual bool ClosedByPlayer(cPlayer & a_Player, bool a_CanRefuse) override;
-	virtual void OpenedByPlayer(cPlayer & a_Player) override;
-	
-protected:
-	cWorld * m_World;
-	int m_BlockX, m_BlockY, m_BlockZ;  // Position of the chest, for the window-close packet
-	cChestEntity * m_PrimaryChest;
-	cChestEntity * m_SecondaryChest;
-} ;
-
-
-
-
-
-class cMinecartWithChestWindow :
-	public cWindow
-{
-public:
-	cMinecartWithChestWindow(cMinecartWithChest * a_ChestCart);
-	~cMinecartWithChestWindow();
-private:
-	cMinecartWithChest * m_ChestCart;
-};
-
-
-
-
-
-class cEnderChestWindow :
-	public cWindow
-{
-public:
-	cEnderChestWindow(cEnderChestEntity * a_EnderChest);
-	~cEnderChestWindow();
-
-protected:
-	cWorld * m_World;
-	int m_BlockX, m_BlockY, m_BlockZ;  // Position of the enderchest, for the window-close packet
-};
-
-
-
-
-
-class cInventoryWindow :
-	public cWindow
-{
-public:
-	cInventoryWindow(cPlayer & a_Player);
-	
-protected:
-	cPlayer & m_Player;
-
-} ;
-
